Fix powertrain maps to address a quirk in the current ChSimpleMapPowertrain implementation.
If the powertrain maps are such that they give exactly zero torque at zero speed, it's possible to get stuck and not be able to move the vehicle.
This can occur if the shaft speed has a negative value and we are in a forward gear (in which case the calculated motor speed is clamped to 0).