replace a part of the mesh broadphase CD with implementation carried out on the device, via CUB
The implementation is barely faster, but it avoids implementing part of the solution on the host,
part on the device. Right now, the solution is almost entirely based on CUB. The way things are
set up, the code is expected to deal with fine meshes better, since CUB is really fast. It remains
to see if this is the case.