Adding templatized versions of 5 ANCF elements with optimizations for execution speed with the option for a “continuous integration” or “pre-integration” method for the generalized internal force and its Jacobian.  Breakout ANCF Materials into standalone classes.
Added the following 5 ANCF elements:
 - Fully parameterized 2 node beam 3243 (ChElementBeamANCF_3243)
 - Gradient deficient 3 node beam 3333 (ChElementBeamANCF_3333)
 - Fully parameterized 4 node shell 3443 (ChElementShellANCF_3443)
 - Higher order 8 node shell 3833 (ChElementShellANCF_3833)
 - Fully parameterized 8 node brick 3843 (ChElementBrickANCF_3843)

Each element is templatized by the number of Gauss quadrature points used for the generalized internal force vector calculation and its Jacobian.  Default parameters are set for each element that yielded a good balance between performance and accuracy compared to a full number of Gauss quadrature points for each element.

Each element can be set to use either a “continuous integration” or the Liu et al. based “pre-integration” calculation method for the generalized internal force vector.  In most cases, the default “continuous integration” method will likely be faster.  If using shell elements with many discrete layers, then the “pre-integration” method could be faster.  Testing has shown that ChElementBeamANCF_3333 is significantly faster than the existing equivalent ChElementBeamANCF element and ChElementShellANCF_3833 is significantly faster than the existing equivalent ChElementShell_8 element.  As a result, ChElementBeamANCF and ChElementShell_8 might be removed in the future since they are redundant.

Details on the generalized internal force and Jacobian calculations can be found in UW-Madison SBEL Tech Report TR-2020-09 (to be published) <https://sbel.wisc.edu/technicalreports/>

Modified ChElementBeamANCF to use the new standalone ChMatrialBeamANCF class.  This class has code specific to the ChElementBeamANCF which will be removed if ChElementBeamANCF is removed in the future.