Fix bugs in ChLoad:
1.
ChLoad<>::LoadIntLoadResidual_F() did not take into consideration loadable variables that may be inactive (e.g. because a node of an FEA is set fixed).
Same issue in ChLoadCustom::LoadIntLoadResidual_F().
The fix is based on a new virtual method (IsSubBlockActive) which must return a boolean indicating whether or not the specified sub-block of DOFs of a loadable element is active or not.

2.
The same mechanism, based on IsSubBlockActive, was implemented in ChLoadCustomMultiple::LoadIntLoadResidual_F().
The previous method, based on testing whether or not an object of type ChVariable was active or not was incorrect.  Indeed, a loadable can have more than one ChVariable objects per sub-block!