Modify FEA bookkeeping to take into account nodes and elements with potentially fixed DOFs.
Note that implementation of the various FEA element functions still assumes that containers (vector and matrices) sized for the full set of element DOFs are provided.
All bookkeeping is done at the mesh level.  While this may be slightly less efficient than requesting each concrete FEA element to load minimum-size vectors and matrices, it simplifies implementation of new elements which need not care about potentially fixed DOFs.
The assumption is that all active (not fixed) DOFs of an FEA node are contiguous at the beginning of the node variable list.