Change definition of box container shapes for collision and proximity detection.
A box container is defined as a box volume delimited by walls of given dimensions (and thickness).
A container is defined relative to a specified frame such that the center of the interior volume is at the frame origin and the box axes aligned with the frame directions.

Both utility functions chrono::utils::AddBoxContainer and chrono::fsi::AddBoxContainerBCE follow the same cnvention.
(Note that chrono::fsi::AddContainerBCE was renamed to chrono::fsi::AddBoxContainerBCE for consistency)