URDF parser: transfer visualization and collision shapes from discarded bodies to parent body.
URDF links with zero mass or inertia are not converted to Chrono bodies.  These links are allowed as a convenience to carry separate visualization and/or collision shapes (they must be connected through a fixed joint to their parent).   The Chrono URDF parser moves all visualization and collision shapes defined for such a link to its parent (taking into account the additional parent-link frame transform).