Add simple models of hydraulic circuit components and hydraulic actuators.
These are based on the lumped fluid theory.
Models for a hydraulic cylinder, a directional valve, and a throttle valve are currently implemented.
Two models of hydraulic actuators, one using two volumes and the other three volumes, are provided.

The hydraulic actuator models are implemented as physcis items with external dynamics and are modeled with ODEs describing the time evolution of the valve input and of the various pressures in the circuit.

The hydraulic actuators can be used stand-alone, for example in a co-simulation setup (in which case they receive the current actuator length and rate of change and return a force) or else can be attached between to Chrono bodies.